+++ react/Image.tsx
@@ -22,6 +22,7 @@ export interface ImageProps
   minHeight?: string | number
   blockClass?: string
   experimentalPreventLayoutShift?: boolean
+  experimentalSetExplicitDimensions?: boolean
   classes?: CssHandlesTypes.CustomClasses<typeof CSS_HANDLES>
   preload?: boolean
   /**
@@ -85,6 +86,7 @@ function Image(props: ImageProps) {
     link,
     title,
     experimentalPreventLayoutShift,
+    experimentalSetExplicitDimensions,
     analyticsProperties = 'none',
     promotionId,
     promotionName,
@@ -117,6 +119,16 @@ function Image(props: ImageProps) {
 
   const placeholderSize = height ?? minHeight ?? maxHeight ?? 'auto'
 
+  const widthWithoutUnits = width ? width.toString().replace(/\D/g, '') : null
+  const heightWithoutUnits = height
+    ? height.toString().replace(/\D/g, '')
+    : null
+
+  const explicitDimensionsAreAvailable =
+    !width?.toString().includes('%') &&
+    !height?.toString().includes('%') &&
+    (widthWithoutUnits || heightWithoutUnits)
+
   const { isMobile } = useDevice()
   const responsiveSrc = isMobile && mobileSrc ? mobileSrc : src
 
@@ -133,6 +145,12 @@ function Image(props: ImageProps) {
       style={imageDimensions}
       ref={imageRef}
       className={handles.imageElement}
+      {...(experimentalSetExplicitDimensions && explicitDimensionsAreAvailable
+        ? {
+            width: widthWithoutUnits ?? undefined,
+            height: heightWithoutUnits ?? undefined,
+          }
+        : {})}
       {...(preload
         ? {
             'data-vtex-preload': 'true',